Change resolution_rule_complete in apply_resolution_rule and test
Has been changed the logic behind this function, because there is no reason to apply the
resolution rule more than one time on the same pair of clauses.
This is because if the resolution rule can be applied for example two times, it means that
there are at least two distinct couple of complementary literals in the two clauses.
When the resolution rule is applied to one these complementary literals in the resulting
clause there will be the other two complementary literals so the clause can be removed
because it is a tautology. The same result is obtained when the resolution rule is applied
on the other couple of complementary literals.
This theory needs to be verified.